+++ plugins/resources/structfile/libstructfile.cpp
@@ -1151,7 +1151,7 @@ extern "C" {
     // interface for POSIX lseek
     eirods::error tar_file_lseek_plugin( 
         eirods::resource_operation_context* _ctx, 
-        size_t                              _offset, 
+        long long                           _offset, 
         int                                 _whence ) {
         // =-=-=-=-=-=-=-
         // check incoming parameters
